// src/interfaces/IMessages.ts
export enum IMessages {
    OperationSuccess = "İşlem Başarılı",
    OperationFailed = "İşlem Başarısız",
    TokenFailed = "Token Geçersiz",
    AuthenticationFailed = "Authentication Failure",
    NotFound = "Not Found",
    Forbidden = "Forbidden",
    BadParameters = "Bad Parameters",
    InternalError = "Internal Error",
    AlreadyRegistered = "Mail Adresi Kayıtlı",
    FillRequiredFields = "Zorunlu alanları doldurunuz"
}